+++ sewer/tests/test_crypto.py
@@ -11,19 +11,32 @@ KeyType = Tuple[str, int]
 class Test_AcmeKey(unittest.TestCase):
 
     rsa_key_types = (("rsa2048", 2048), ("rsa3072", 3072), ("rsa4096", 4096))
-    secp_key_types = (("secp256r1", 256), ("secp384r1", 384), ("secp521r1", 521))
+    secp_key_types = (("secp256r1", 256), ("secp384r1", 384))
 
     def fromfile_privbytes_frombytes_sign_key(self, key_type: KeyType) -> None:
+        """
+        this has grown into a test for almost everything we do with keys
+        """
+
         type_name, key_size = key_type
         filename = "test/%s.pem" % type_name
+
         # from_file
         loaded_key = AcmeKey.from_file(filename)
         self.assertTrue(loaded_key.pk.key_size == key_size)
-        # private_bytes
+
+        # private_bytes ("assert" no exception)
         loaded_pb = loaded_key.private_bytes()
+
+        # private bytes matches original file
+        with open(filename, "rb") as f:
+            file_bytes = f.read()
+        self.assertTrue(loaded_pb == file_bytes)
+
         # from_bytes
         reloaded_key = AcmeKey.from_bytes(loaded_pb)
         self.assertTrue(loaded_key.pk.private_numbers() == reloaded_key.pk.private_numbers())
+
         # sign_message
         self.assertTrue(len(loaded_key.sign_message(b"Taketh uth to thine leaderth")))